Refactor BrowserTab and Backend to improve lastAccessed handling
Updated the lastAccessed property in the Backend to ensure it is always a double. Refactored the _BrowserTab widget to display the last accessed time, enhancing the user interface with additional tab information. This change improves the overall functionality and user experience in managing browser tabs.

+++ lib/store/backend.dart
@@ -175,7 +175,7 @@ extension _$Backend on _Backend {
             url: url,
             title: title,
             windowId: windowId ?? -1,
-            lastAccessed: lastAccessed ?? -1.0,
+            lastAccessed: (lastAccessed ?? -1.0).toDouble(),
           );
         case "tab_size_change":
           final tab = HF.json(json["tab"]);
